third_party/blink/web_tests/external/wpt/notifications/cross-origin-nested.tentative.https.sub-expected.txt

This is a testharness.js-based test.
[FAIL] third party iframe
  assert_equals: should deny the permission expected "denied" but got "granted"
[FAIL] nested first party iframe
  assert_equals: should deny the permission expected "denied" but got "granted"
[FAIL] nested first party worker
  assert_equals: should deny the permission expected "denied" but got "granted"
Harness: the test ran to completion.
